LPD-58182 Inline BooleanQueryTranslator/BooleanQueryTranslatorImpl
