--- modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/test/java/com/liferay/portal/search/elasticsearch7/internal/query/ElasticsearchQueryTranslatorFixture.java
@@ -1,28 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.portal.search.elasticsearch7.internal.query;
-
-import com.liferay.portal.kernel.test.ReflectionTestUtil;
-
-/**
- * @author Michael C. Han
- */
-public class ElasticsearchQueryTranslatorFixture {
-
-	public ElasticsearchQueryTranslatorFixture() {
-		ReflectionTestUtil.setFieldValue(
-			_elasticsearchQueryTranslator, "_booleanQueryTranslator",
-			new BooleanQueryTranslatorImpl());
-	}
-
-	public ElasticsearchQueryTranslator getElasticsearchQueryTranslator() {
-		return _elasticsearchQueryTranslator;
-	}
-
-	private final ElasticsearchQueryTranslator _elasticsearchQueryTranslator =
-		new ElasticsearchQueryTranslator();
-
-}
